+++ app/src/main/java/ru/practicum/android/diploma/presentation/ui/root/RootActivity.kt
@@ -2,20 +2,24 @@ package ru.practicum.android.diploma.presentation.ui.root
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
-import ru.practicum.android.diploma.BuildConfig
+import androidx.navigation.fragment.NavHostFragment
+import androidx.navigation.ui.setupWithNavController
+import com.google.android.material.bottomnavigation.BottomNavigationView
 import ru.practicum.android.diploma.R
 
 class RootActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
+        // 1. Устанавливаем XML разметку
         setContentView(R.layout.activity_root)
 
-        // Пример использования access token для HeadHunter API
-        networkRequestExample(accessToken = BuildConfig.API_ACCESS_TOKEN)
-    }
+        // 2. Инициализируем навигацию
+        val navHostFragment = supportFragmentManager
+            .findFragmentById(R.id.rootFragmentContainerView) as NavHostFragment
+        val navController = navHostFragment.navController
 
-    private fun networkRequestExample(accessToken: String) {
-        // ...
+        // 3. Связываем BottomNavigationView с контроллером
+        val bottomNav = findViewById<BottomNavigationView>(R.id.bottomNavigationView)
+        bottomNav.setupWithNavController(navController)
     }
-
 }
