--- app/src/main/java/ru/practicum/android/diploma/ui/root/RootActivity.kt.kt
@@ -1,25 +1,0 @@
-package ru.practicum.android.diploma.ui.root
-
-import android.os.Bundle
-import androidx.appcompat.app.AppCompatActivity
-import androidx.navigation.fragment.NavHostFragment
-import androidx.navigation.ui.setupWithNavController
-import com.google.android.material.bottomnavigation.BottomNavigationView
-import ru.practicum.android.diploma.R
-
-class RootActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        // 1. Устанавливаем XML разметку
-        setContentView(R.layout.activity_root)
-
-        // 2. Инициализируем навигацию
-        val navHostFragment = supportFragmentManager
-            .findFragmentById(R.id.rootFragmentContainerView) as NavHostFragment
-        val navController = navHostFragment.navController
-
-        // 3. Связываем BottomNavigationView с контроллером
-        val bottomNav = findViewById<BottomNavigationView>(R.id.bottomNavigationView)
-        bottomNav.setupWithNavController(navController)
-    }
-}
