+++ app/src/main/res/color/bottom_nav_colors.xml
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="utf-8"?>
+<selector xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <item
+        android:color="@color/blue"
+        android:state_checked="true" />
+
+    <item
+        android:color="@color/gray" />
+
+</selector>
