+++ app/src/main/res/layout/activity_root.xml
@@ -20,7 +20,17 @@
         android:id="@+id/bottomNavigationView"
         android:layout_width="0dp"
         android:layout_height="wrap_content"
+
+        android:background="@color/white"
+
+        app:itemIconTint="@color/bottom_nav_colors"
+        app:itemTextColor="@color/bottom_nav_colors"
+
+        app:itemActiveIndicatorStyle="@null"
+        app:itemRippleColor="@null"
+
         app:menu="@menu/bottom_navigation_view"
+
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintEnd_toEndOf="parent" />
