+++ app/src/main/res/menu/bottom_navigation_view.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <menu xmlns:android="http://schemas.android.com/apk/res/android">
     <item
-        android:id="@+id/mainFragment"
+        android:id="@+id/searchFragment"
         android:icon="@drawable/ic_main_24"
         android:title="@string/tab_main" />
     <item
