+++ app/src/main/res/navigation/nav_graph.xml
@@ -2,21 +2,21 @@
 <navigation xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:id="@+id/nav_graph"
-    app:startDestination="@id/mainFragment">
+    app:startDestination="@id/searchFragment">
 
     <fragment
-        android:id="@+id/mainFragment"
-        android:name="ru.practicum.android.diploma.ui.screens.MainFragment"
+        android:id="@+id/searchFragment"
+        android:name="ru.practicum.android.diploma.presentation.search.SearchFragment"
         android:label="Search" />
 
     <fragment
         android:id="@+id/favoritesFragment"
-        android:name="ru.practicum.android.diploma.ui.screens.FavoritesFragment"
+        android:name="ru.practicum.android.diploma.presentation.favorites.FavoritesFragment"
         android:label="Favorites" />
 
     <fragment
         android:id="@+id/teamFragment"
-        android:name="ru.practicum.android.diploma.ui.screens.TeamFragment"
+        android:name="ru.practicum.android.diploma.presentation.team.TeamFragment"
         android:label="Team" />
 
 </navigation>
